+++ Gameplay/Mods/Mirror.cs
@@ -20,14 +20,14 @@ namespace YAVSRG.Gameplay.Mods
             }
         }
 
-        private int BitMirror(int v, int k)
+        private ushort BitMirror(ushort v, int k)
         {
-            int o = 0;
+            ushort o = 0;
             for (int i = 0; i < k; i++)
             {
                 if ((1 << i & v) > 0)
                 {
-                    o += (1 << (k - 1 - i));
+                    o += (ushort)(1 << (k - 1 - i));
                 }
             }
             return o;
